# specs/002-authentication/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Authentication API
  description: User registration and login with JWT authentication
  version: 1.0.0
  contact:
    name: Hackathon Todo API
    url: http://localhost:8000

servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.yourdomain.com/api/v1
    description: Production server (future)

tags:
  - name: auth
    description: Authentication endpoints

paths:
  /auth/signup:
    post:
      tags:
        - auth
      summary: Register new user
      description: |
        Create a new user account with email and password.
        Returns JWT token for immediate authentication.
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            examples:
              validUser:
                summary: Valid registration
                value:
                  email: "user@example.com"
                  password: "securepassword123"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                success:
                  summary: Successful registration
                  value:
                    access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1NTBlODQwMC1lMjliLTQxZDQtYTcxNi00NDY2NTU0NDAwMDAiLCJleHAiOjE3MzU2ODk2MDAsImlhdCI6MTczNTY4NjAwMH0.signature"
                    token_type: "bearer"
                    user:
                      id: "550e8400-e29b-41d4-a716-446655440000"
                      email: "user@example.com"
                      created_at: "2025-01-01T00:00:00Z"
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidEmail:
                  summary: Invalid email format
                  value:
                    detail: "Invalid email format"
                shortPassword:
                  summary: Password too short
                  value:
                    detail: "Password must be at least 8 characters"
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicate:
                  summary: Duplicate email
                  value:
                    detail: "Email already registered"
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  /auth/login:
    post:
      tags:
        - auth
      summary: Login existing user
      description: |
        Authenticate with email and password.
        Returns JWT token on successful authentication.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              validLogin:
                summary: Valid credentials
                value:
                  email: "user@example.com"
                  password: "securepassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                success:
                  summary: Successful login
                  value:
                    access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1NTBlODQwMC1lMjliLTQxZDQtYTcxNi00NDY2NTU0NDAwMDAiLCJleHAiOjE3MzU2ODk2MDAsImlhdCI6MTczNTY4NjAwMH0.signature"
                    token_type: "bearer"
                    user:
                      id: "550e8400-e29b-41d4-a716-446655440000"
                      email: "user@example.com"
                      created_at: "2025-01-01T00:00:00Z"
        '400':
          description: Missing email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingFields:
                  summary: Missing required fields
                  value:
                    detail: "Email and password are required"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidCreds:
                  summary: Invalid email or password
                  value:
                    detail: "Invalid email or password"
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

components:
  schemas:
    SignupRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: Valid email address
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 128
          format: password
          description: Password (8-128 characters)
          example: "securepassword123"

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: "user@example.com"
        password:
          type: string
          format: password
          description: User password
          example: "securepassword123"

    UserResponse:
      type: object
      required:
        - id
        - email
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: User UUID
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          description: User email address
          example: "user@example.com"
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp (ISO 8601)
          example: "2025-01-01T00:00:00Z"

    AuthResponse:
      type: object
      required:
        - access_token
        - token_type
        - user
      properties:
        access_token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          enum: ["bearer"]
          default: "bearer"
          description: Token type (always 'bearer')
        user:
          $ref: '#/components/schemas/UserResponse'

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Email already registered"

    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
                description: Location of validation error
              msg:
                type: string
                description: Validation error message
              type:
                type: string
                description: Error type
          example:
            - loc: ["body", "email"]
              msg: "value is not a valid email address"
              type: "value_error.email"

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from /auth/signup or /auth/login.
        Include in Authorization header as: "Bearer <token>"

security: []  # Auth endpoints don't require authentication